Use an unit type wrapper for Piece
The motivation is when I'm working on the issue of deduplicating the RPC types,
I found that `RpcArchivedSegment` can be improved, but currently serde
only supports the arrays up to 32 elements, therefore a type wrapper for
the big array is neccessary.

There are two crates for the big array serde at the moment:

- https://github.com/est31/serde-big-array
- https://github.com/Kromey/serde_arrays

No particular reason, I choose serde_arrays just because it supports
nested array while serde-big-array does not support yet.