Fix parachain stalled at #2
The root cause is that the parachain chain block production and import
are performed **concurrently**, but the correct block import is order
sensitive, which means the executor node can't produce next candidate
until the current candiate has been imported, otherwise it'll produce
multiple candidates at the same height. This implies that the logic
around `check_block_status` is crucial.