Introduce OpaqueBundle and integrate tx pool in bundle production
`OpaqueBundle` with opaque extrinsics will be used on primary chain
as the primary chain does not care about the extrinsic format, it can be
evolved as `CompactBundle` or something like that in the future.

This PR also starts an initial transaction pool integration in bundle
production, not the final version, but should be a step towards it.