Set the default ss58 version on node start up
The default 558 version only impacts the address form when you
convert an AccountId to an Address, e.g., if you have an RPC whose
return value contains a field with type `AccountId`, this field can
be ss58 encoded using an unexpected address format if the global default
version(`SubstrateAccount` by default) is not set properly.

This basically won't lead to an error anywhere, just the output address from the node can be wrong.