Store Execution Receipt Keyed by block hash
Store the execution receipt by block hash, use another storage for
tracking the associated block number to have knowledge of if it's too
old and should be pruned.

This commit also removes the deletion of local ER even it's correctly
validated, now the obsolete ERs will be removed soly according to the
prune depth.