Introduce `SecondaryApi_apply_extrinsic_with_post_state_root`
Adding an idential API of `apply_extrinsic` but call the storage root
at the end and use it as the return value. And then we can interprete
the return value as `post_state_root` and compare it with the one
received from the network.

Initially an idea from https://github.com/paritytech/substrate/pull/10922#issuecomment-1066752530