Ensure no primary blocks will be missed on the executor side
The imported block notifications and overseer config initialization
happen concurrently and we only process the blocks when the config is
`Some`, casuing dozens of blocks that was imported during the config
initialization period can be ignored. This patch aims to solve it by
caching these blocks and process them once the config has been
initialized. The future solution can be initialize the config when we
create the Overseer, but this requires a bit more refactorings.