Skip the signed execution receipt processing if major syncing
The new slot notification is disabled while the primray chain is major
syncing, which means the bundle production is disable in the meantime.
Similarly, I think it makes sense to also skip the signed execution
receipt processing while major syncing. Assuming the executor chain
is built by only processing the primary blocks, then we can use the
network sync state of primary chain as an indicator to skip that.