Skip processing the genesis primary block on executor
Since
https://github.com/subspace/subspace/commit/2fc664a4ad2974bd81322623775d592c997046cf
the primary genesis block is also processed by executor on startup,
resulting an issue that the block number of execution chain is always ahead of the
consensus chain block number by 1. The reason is that previously the
executor only processes the primary block after the config has been
intialized, hense the active leaves processing before the loop will be
actually skipped, but now the leaves will always be processed as the
config is ensured to be initialized.

For this specific case, it makes sense to me that there is no leaves for
the genesis block.