Fix `execute_block` of cirrus-pallet-executive
The sympotom is that previously when we sync a block from the other
executors, an error `storage root does not match` will occur, but the
execution chain is still able to proceed due to we also try to produce a
block based on the primary block in the meanwhile. This patch simply
fixes the syncing issue from other executors.

Basically, we have two approaches to make the execution chain moving
forward:

- Create new execution chain block based on the primary block
- Sync the block from the other executors.

https://github.com/paritytech/substrate/blob/d26123aec44e3df00b921ab0d27cbc1de0b11fa9/client/service/src/client/client.rs#L816
Runtime api `execute_block` is used to process the block synced from the other
executors, while the block produced using the primary block uses the
custom BlockBuilder to execute the block(`apply_extrinsic` +
`finalize_block`). `execute_block` is essentially a combination of
`apply_extrinsic` and `finalize_block`, but the execution trace
recording that was added to `finalize_block` was missing in
`execute_block` before.