Bring down the whole service on bundles processor error
If there is an error occurred during the primary block processing on the
executor side, take down the whole service as bundles processor is
considered as an essential task.

Also add a TODO of adding a block import throttling to signify a primary
block has been processed by the executor.