Throttle the primary block import when bundles processor is slow
If the executor's bundles processor is much slower than the embedded
primary node syncing, we should hold off the primary node syncing until
the bundles processor catches up, otherwise the primary block can be
pruned but not yet to be processed by the executor since the embedded
primary node is not expected to be an archival node.

In order to hold off the primary node syncing, we take the way of
holding off the block import with a channel, if the channel is full,
which means there are too many pending blocks waiting to be processed by
the executor, the block import will just await, which effitively stops
the syncing.