Receipts can be empty for the first block
There is an exception that for the bundles created during period of best
block is the first block, the receipts can be empty as the genesis
receipt is already initialized on the primary chain.

On the executor side, except for the genesis primary block, the receipts
can not be empty for the consequent blocks.