Verify the state root in the core domain bundle
`core_block_hash` and `core_state_root` are added to the struct
`ProofOfElection` as we need to the block hash and the corresponding
state root of a core domain. They are optional for the sytem domain
bundle.

Also added a TODO to rename `secondary_hash` to `domain_hash` in
the Receipt struct since now the core domain is possible as well.