Fix core domain state root verification
This commit fixes the bug that the system block number was misused as the
core block number. I was likely drunk while initially writing this.
I ran a few hunreds of blocks locally before but wasn't able to catch it
earlier probably due to the coincidence that the system block number happens
to match the core domain block number. After the recent consensus change, it
became to be more obvious causing the error `StateRootUnveriable` with
slower farming.

Having a look at the system domain state root verification in
pallet-domains can also help understand the changes.