Fix the domain block notification using `BlockchainEvents`
Previously, we added the block notification stream in `import_queue` of
domain client, which does not work in fact we don't import domain blocks
using the import queue and instead import the domain block using the bare
`BlockImport` implemented by the client.

This commit fixes it by making use of the `BlockchainEvents` provided by
the client itself, it also simplifies the code a lot, the copy-pasted
notification module can be removed now.