Remove incorrect assertion on primary block number while processing new domain block
This assertion essentially expects the best number of domain chain increases
strictly by 1, which is incorrect because the domain chain is entirely driven
by the primary chain and primary chain fork is legally possible.

This commit removes it as the domain already follows the fork choice of
primary chain, that being said, the domain chain forks when the primary chain fork occurs.