Check Receipt::UnknownBlock of core domain bundle in pallet-domain-registry
What this commit does:

- pallet-domains: only check ReeiptError::UnknownBlock for system
  domain bundle and remove the check for core domain bundle.
- pallet-domain-registry: check ReceiptError::UnknownBlock for core
  domain receipts in `pre_dispatch_submit_core_bundle()`, using the
  `BlockHash` extended in the system domain header digest.

Previously, it's the primary chain that ensures the receipt of all
domains points to a valid primary block, which is problematic in this
scenario: when a core domain stalls, the system domain moves forward as
normal, once the receipt is deep enough, the corresponding primary block
info `BlockHash` will be pruned along with the receipt pruning. Then
when the core domain resumes, pallet-domains can not verify if the core
domain receipt points to a valid primary block. The solution is to
maintain a copy of `BlockHash` for each domain in
pallet-domain-registry.