Ensure receipts are consecutive on primary chain
The range of receipts in a core bundle on gemini-3b is [..., 514, 516],
515 is somehow skipped, I have no idea why such receipts can be possible
and still exploring the explanation.

It does not corrupt the state as the bundle will be detected as invalid
while checking the missing receipt, but it's obvious that there is a glitch
here: we should check if the receipts are consecutive instead of whether it's
simply sorted as `is_sorted` does not gurantee the consecution, i.e., no
missing receipt.