Verify `pre_state_root` in InvalidStateTransitionProof
In order to verify `pre_state_root`, we need to retrieve the full trace
of a specific receipt, and then compare it with the one at the index
specified by the execution phase.

`ExecutionPhase` will be refactored to include the trace index only
instead of shipping the full call data later.