define state root extractor and provide an impl with system domain client.
This is the main crust of the PR. Beacuse we want to use XDM verifier during the fraud proof verification on primary chain client but node may or may not be running the system domain which is currently used to extract state roots using System domain runtime. We provide a trait bound to extract the state roots so that we can inject system domain client or other implementation from primary chain.