provides an implementation of the StateRootExtractor with just system domain runtime WASM.
This implementation can be used on primary chain while doing fraud proof verification. The runtime code is expected as argument since we would have extracted the runtime from primary chain for fraud proof anyway. As for the executor, we can reuse the exisiting one on the primary chain.

Note: this implementation will be used in next PRs.