update StateRootExtractor to pass block hash from the caller.
When we use syatem domain client to extract state roots when verifying TXN using TX pool and Bundleverifier which are assumed to use best hash of the chain. When we use system domain runtime, we expect caller to provide a valid hash which is used to decode the extrinsic.

Block hash is only used to setup the right backend and is not used for anything when runtime is used.