Rename `primary_hash` to `primary_parent_hash` and fix a bug when generating the invalid state transition proof
The main point of this commit is to fix a bug that the legacy `parent_hash`
in `InvalidStateTransitionProof` was incorrectly initialized as the domain
parent hash, which should be the parent hash of the parent block(used to
retrieve the domain runtime code from the primary chain using the
primary chain runtime api).

We didn't catch it earlier because it was
manually set to the _right_ hash in the test by accident which is essentially
caused the gap between the test and production environment. I hope it
won't happen again when we add tests closer to the production code with
the new testing framework.

This commit also makes it more explicit by renaming `parent_hash` to `primary_parent_hash`.