Introduce `PreValidateTransaction` and merge generics `Verifier` and `BundleValidator` into `TxPreValidator`
This commits introduces a trait `PreValidateTransaction` which is
invoked before calling the normal `validate_transaction` function.

`FraudProofVerifierAndBundleValidator` is separated out temperoarily,
it's now used by both the primary chain and the system domain. However,
the system domain does not need the bundle validator at all, which will
be removed later, making it more explicit and easier to figure out the
ingredients of the custom tx-pre-validator for primary_chain/system_domain/core_domain.