Track the mapping of (domain_hash, primary_hash) in aux db
Upon the incoming primary block notification, we need the hash of
primary block that derives the current domain block in order to
calculate the primary chain tree route.

It was retrieved from the receipt related to the best domain block
before, now this mapping has to be stored separately in the aux db
due to the change of receipts tracking, it will be pruned periodically
later.