Introduce produce_blocks macro to replace futures::join in test
When using futures::join if the primary node fail to produce block the whole future will stuck
because the domain node will also not able to produce block, which hurt debug efficiency a lot.

This commit also come with the produce_blocks_until macro which will be used in the upcome test.

Signed-off-by: linning <linningde25@gmail.com>