Add necessary APIs for `ReceiptsChecker` in ParentChainInterface
The following APIs are needed in ReceiptsChecker.

- `block_body`
- `extract_receipts`
- `extract_fraud_proofs`
- `oldest_receipt_number`

This commit also changes the trait `SystemDomainApi` by adding a new
generic `CHash` (which is not a good name, stands for the hash type of
the other domains the system domain secures). This change is primarily
to make the extraction in ReceiptsChecker easier. Since we already
assume `NumberFor<Block>`/`NumberFor<SBlock>`/`NumberFor<PBlock>` can be
mutually transformed, so it technically should be fine.