use balance hold mechanism instead of freeze, fix TODOs related to treasury account, and cleanup tests
Unfortuantely, freeze mechanism does not work like we intended. If there are multiple Freezes, then only maximum of those is considered frozen. The problem became apparent when we introduced multiple freeze ids for the staking. Turns out Hold mechanism is the way to go and seems to be built for Staking. This integration cleaned some code but penalised on other sections.