Remove the bundle validator
The bundle validator is mostly used to perform the duplicated bundle check
but now this check is more nature to perform at the runtime since we need
to track the inboxed bundle there. There are also some other misc check in
the bundle validator but these check is deplicated with the validate_bundle
runtime check thus we can safely remove them.

Signed-off-by: linning <linningde25@gmail.com>