derive correct genesis domain hash and correct genesis er.
This commit adds domain hash derivation given required info for header. Since Digest is always empty for Genesis block, we can already derive the genesis domain hash. For the rest of the domain blocks, we would need to do a fraud proof that can provide the digests pointing to the state root of the domain block.